+++ fj-daogen-base/src/main/java/org/fugerit/java/daogen/base/gen/Junit4ModelGenerator.java
@@ -30,7 +30,8 @@ public class Junit4ModelGenerator extends DaogenBasicGenerator {
 		this.setJavaStyle( STYLE_CLASS );
 		GenUtils.addAll( this.getImportList(), "org.slf4j.Logger", "org.slf4j.LoggerFactory", "org.junit.Assert", "org.junit.Test",
 				this.getDaogenConfig().getGeneralProp( DaogenCatalogConstants.GEN_PROP_PACKAGE_MODEL )+"."+this.getEntityModelName(),
-				this.getDaogenConfig().getGeneralProp( DaogenCatalogConstants.GEN_PROP_PACKAGE_HELPER )+"."+this.getEntityHelperName());
+				this.getDaogenConfig().getGeneralProp( DaogenCatalogConstants.GEN_PROP_PACKAGE_HELPER )+"."+this.getEntityHelperName(),
+				this.getDaogenConfig().getGeneralProp( DaogenCatalogConstants.GEN_PROP_PACKAGE_HELPER )+"."+this.getEntityWrapperName() );
 	}
 
 	@Override